+++ www/sites/activatemail.php
@@ -7,10 +7,19 @@ list($UUID, $EMAIL) = $DbLink->next_record();
 }
 
 if($UUID)
-{
-	$WERROR="Thank you, your email address was changed";
-	$DbLink->query("UPDATE ".C_USERS_TBL." SET email='".$EMAIL."' WHERE UUID='".$UUID."' ");
-	$DbLink->query("DELETE FROM ".C_CODES_TBL." WHERE code='$_GET[code]' and info='confirm'");
+{	
+	$found = array();
+	$found[0] = json_encode(array('Method' => 'SaveEmail', 'WebPassword' => md5(WIREDUX_PASSWORD)
+		, 'UUID' => $_SESSION[USERID]
+		, 'Email' => $EMAIL));
+	$do_post_requested = do_post_request($found);
+	$recieved = json_decode($do_post_requested);
+	
+	if ($recieved->{'Verified'} == "true") 
+	{
+		$WERROR="Thank you, your email address was changed";		
+		$DbLink->query("DELETE FROM ".C_CODES_TBL." WHERE code='$_GET[code]' and info='confirm'");
+	}
 }
 else
 {
